refactor(automations): the trigger notifier joins the server package; delivery tier complete (ADR-063 P1)
TriggerNotifier (the test-fire delivery contract) moves out of
trigger-template.service into dispatch/notifier alongside
createLiveTriggerNotifier({ mailer, webhookSender, slackWebApi }) — the
mailer becomes the package's MailerPort, the other two ports are the
package's own client types. The app composes delivery/appTriggerNotifier
from its mailer and configured senders. src/server app-layer delivery/ now
holds only the app adapters and the real-validator tests.
